src/requirements.txt
weaviate-client
python-dotenv
requests